Mapping by TFID,TLID now instead of TFID,fullname
The reason for this, is that different address ranges with the same
fullname can have different gps coordinates associated with those
particular address ranges. So as per TIGER documentation we should use
TFID,TLID pairs for more accurate data.  The output will now have
several records with the same street names and that’s fine, because
their TLIDs are different, meaning they correspond to different GPS
coordinates in the database.